requirements.txt
# Core dependencies
Flask==2.3.3
openai==1.6.0
python-dotenv==1.0.0
Werkzeug==2.3.7
jinja2==3.1.2
itsdangerous==2.1.2
click==8.1.7

# Database
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
SQLAlchemy==2.0.23

# Web scraping and content extraction
beautifulsoup4==4.12.2
requests==2.31.0
trafilatura==1.6.2
newspaper3k==0.2.8
readability-lxml
langdetect==1.0.9

# Audio processing
pydub==0.25.1

# NLTK for natural language processing
nltk==3.8.1

# Rate limiting and security
Flask-Limiter==3.5.0
limits==3.6.0

# Caching
Flask-Caching==2.1.0
redis==5.0.1

# Production server
gunicorn==21.2.0
gevent==23.9.1

# Utilities
PyJWT==2.8.0
bcrypt==4.0.1
email-validator==2.1.0.post1
Markdown==3.5
python-slugify==8.0.1
pytz==2023.3
validators==0.22.0

# Development tools
black==23.9.1
flake8==6.1.0
pytest==7.4.2
pytest-flask==1.2.0